[refactor] like domain
